[한길] [상점시스템] 상점시스템 추가
초기화면에서 3번 누르면 상점에 들어가게 만들었습니다. 2번은 인벤토리로 수정(추가할 예정이라고 합니다)
Shop 클래스를 추가하였고, Item 클래슬를 상속받는 자식클래스 MiniPotion, BigPotion 클래스를 추가하였습니다.
Main함수에서 MiniPotion은 3개, BigPotion은 0개로 초기화하고, 구매하거나 사용하면 해당 포션의 숫자가 변동하게 만들었습니다.
몬스터 클래스의 GivingExp 등 경험치 관련 변수, 함수들 자료형을 통일해야할것 같습니다.

+++ CH2TeamProject/CH2TeamProject/Monster.cpp
@@ -70,7 +70,7 @@ int Monster::GetLvl() const
     return Lvl;
 }
 
-int Monster::GetGivingExp() const  // [한길] 시스템 변경으로 int에서 float으로 변경.
+float Monster::GetGivingExp() const  // [한길] 시스템 변경으로 int에서 float으로 변경.
 {
     return GivingExp;
 }
